requirements.txt
asgiref==3.2.3
astroid==2.3.3
awscli==1.14.8
botocore==1.8.12
colorama==0.4.3
dj-database-url==0.5.0
Django==3.0.3
django-heroku==0.3.1
docutils==0.14
gunicorn==20.0.4
isort==4.3.21
jmespath==0.9.3
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2==2.8.4
pyasn1==0.4.2
pylint==2.4.4
python-dateutil==2.6.1
pytz==2019.3
PyYAML==5.1
rsa==3.4.2
s3transfer==0.1.12
six==1.14.0
sqlparse==0.3.0
typed-ast==1.4.1
virtualenv==16.7.9
whitenoise==5.0.1
wrapt==1.11.2
